refactor: Refactor Service logic and add test cases

+++ src/main/java/com/gdsc/blog/article/service/ArticleService.java
@@ -1,36 +1,56 @@
 package com.gdsc.blog.article.service;
 
+import com.gdsc.blog.article.dto.ArticleCreateDto;
+import com.gdsc.blog.article.dto.ArticleDto;
 import com.gdsc.blog.article.dto.ArticleUpdateDto;
 import com.gdsc.blog.article.entity.Article;
 import com.gdsc.blog.article.repository.ArticleRepository;
+import com.gdsc.blog.mapper.ArticleMapper;
 import com.gdsc.blog.user.entity.User;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import com.gdsc.blog.user.entity.UserRole;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final ArticleMapper articleMapper;
 
     /**
      * 게시글 생성
      *
-     * @param article Article 객체
-     * @param user    작성자
+     * @param articleCreateDto 게시글 생성 DTO
+     * @param user             작성자
      */
-    public Article createArticle(
-        Article article,
-        User user) { //create new article
-        article.setCreateDate(new Timestamp(System.currentTimeMillis()));
-        article.setModifyDate(new Timestamp(System.currentTimeMillis()));
-        article.setUser(user);
-        return articleRepository.save(article);
+    public ArticleDto createArticle(
+            ArticleCreateDto articleCreateDto,
+            User user) {
+
+        // 현재 시간
+        Timestamp timestamp = new Timestamp(System.currentTimeMillis());
+
+        // 게시글 생성
+        Article article = Article.builder()
+                .title(articleCreateDto.getTitle())
+                .content(articleCreateDto.getContent())
+                .user(user)
+                .createDate(timestamp)
+                .modifyDate(timestamp)
+                .build();
+
+        return articleMapper.toDto(articleRepository.save(article));
     }
 
     /**
@@ -38,8 +58,16 @@ public class ArticleService {
      *
      * @return 게시글 목록
      */
-    public List<Article> getUserArticle(User user) { //get all articles
-        return user.getArticles();
+    public List<ArticleDto> getUserArticle(User user) {
+        // get user article list
+        List<Article> articles = articleRepository.findByUser(user);
+        log.info("articles: {}", articles);
+
+        // article list to article dto list
+        List<ArticleDto> articleList = new ArrayList<>();
+        articles.forEach(article -> articleList.add(articleMapper.toDto(article)));
+
+        return articleList;
     }
 
     /**
@@ -48,16 +76,23 @@ public class ArticleService {
      * @param idx 게시글 id
      * @return Article 객체
      */
-    public Article getArticleById(Long idx) { //get article by id
+    public ArticleDto getArticleById(Long idx) { //get article by id
         Optional<Article> article = articleRepository.findById(idx);
 
-        return article.orElseThrow(() -> new NullPointerException("Not found Article by id"));
+        if (article.isPresent()) {
+            return articleMapper.toDto(article.get());
+        } else {
+            throw new IllegalArgumentException("Not found Article by id");
+        }
     }
 
-    public Article getArticleByTitle(String title){
-        Optional<Article> article = articleRepository.findByTitle(title);
+    public List<ArticleDto> findArticleByTitle(String title) {
+        List<Article> articles = articleRepository.findByTitleLike(title);
 
-        return article.orElseThrow(() -> new NullPointerException("Not found Article by title"));
+        List<ArticleDto> articleList = new ArrayList<>();
+        articles.forEach(article -> articleList.add(articleMapper.toDto(article)));
+
+        return articleList;
     }
 
     /**
@@ -68,22 +103,36 @@ public class ArticleService {
      * @param user 작성자
      * @return Article 객체
      */
-    public Article updateArticle(Long id, ArticleUpdateDto dto, User user) {
+    public ArticleDto updateArticle(Long id, ArticleUpdateDto dto, User user) {
 
-        Article article = getArticleById(id); //게시글 id로 게시글 가져오기
+        Article article = articleRepository.findById(id).orElseThrow(); //게시글 id로 게시글 가져오기
+
+        if (!article.getUser().getUsername().equals(user.getUsername())) { //게시글 작성자와 로그인한 유저가 같지 않을 경우
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
+
         article.setTitle(dto.getTitle());
         article.setContent(dto.getContent());
         article.setModifyDate(new Timestamp(System.currentTimeMillis()));
 
-        return articleRepository.save(article);
+        return articleMapper.toDto(articleRepository.save(article));
     }
 
     /**
-     * 게시글 삭제
+     * 게시글 삭제 (관리자, 작성자만 가능)
      *
-     * @param article 게시글 객체
+     * @param idx  게시글 id
+     * @param user 작성자
      */
-    public void deleteArticle(Article article) {
-        articleRepository.delete(article);
+    public void deleteArticle(Long idx, User user) {
+        Article article = articleRepository.findById(idx).orElseThrow(); //게시글 id로 게시글 가져오기
+
+        // 게시글 작성자와 로그인한 유저가 같거나, 로그인 한 유저가 관리자일 경우
+        if (article.getUser().getUsername().equals(user.getUsername()) || user.getRoles().contains(UserRole.ROLE_ADMIN)) {
+            articleRepository.delete(article);
+        } else {
+            throw new NullPointerException("삭제 권한이 없습니다.");
+        }
+
     }
 }
